Move users-repo into separate file. Fixes from code review.
